+++ src/RivaRental.Domain/PriceCalculator.cs
@@ -2,7 +2,7 @@
 
 public class PriceCalculator
 {
-    public static double Calculate(Boat b, int hours, int walkingHours, double extra = 0) =>
+    public static double Calculate(Boat b, int hours, double walkingHours, double extra = 0) =>
         b switch
         {
             Diable diable => diable.CalculatePrice(hours, walkingHours) + extra,
